Add files via upload
First version of the R script using gtrends package with the functionality to extract the hits for one term in a period of 5 years with daily detail. 
Since there is a limitation to extract daily data from gtrends above 8 months, I have use a scale factor for 2 periods with one month in common.